Font added, changed default font, font scaling option added,

+++ src/RomListManager.cpp
@@ -65,7 +65,7 @@ void RomListManager::initText(const Configuration& configuration, const sf::Font
 								configuration.getRom_name_color_alpha()));
 		text.setCharacterSize(configuration.getRom_name_size());
 
-		if(rom.getDescription().length() > configuration.getRom_name_max_length())
+		if((int)rom.getDescription().length() > configuration.getRom_name_max_length())
 		{
 			text.setString(rom.getDescription().substr(0, configuration.getRom_name_max_length()) + configuration.getRom_name_suffix_max_length());
 		}
